dom/tests/mochitest/mouselock/test_mouseLockableHasRequiredMethods.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633602
-->
<head>
  <title>Test for Bug 633602</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js">
  </script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js">
  </script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
  Mozilla Bug 633602</a>
<div id="content">
</div>
  <pre id="test">
    <script type="application/javascript">
      /** Test for Bug 633602 **/
      /** Test to see if MouseLockable has lock(), unlock(), islocked() **/
      SimpleTest.waitForExplicitFinish()
      is(MouseLockable.prototype.hasOwnProperty('lock'), true,
        "MouseLockable does not have lock method.");
      is(MouseLockable.prototype.hasOwnProperty('islocked'), true,
        "MouseLockable does not have islocked method.");
      is(MouseLockable.prototype.hasOwnProperty('unlock'), true,
        "MouseLockable does not have unlock method.");
      SimpleTest.finish();
    </script>
  </pre>
</body>
</html>
